\name{toFile}
\alias{toFile}
\alias{toFile,ANY,character-method}
\alias{toFile,RAutoDocumentation,ANY-method}
\alias{toFile,RTUDocumentation,missing-method}

\title{Write the documentation objects to a connection}
\description{
}
\usage{
toFile(obj, file, ...)
}
\arguments{
  \item{obj}{the documentation object to be written to the connection.}
  \item{file}{the name of a file or a connection object}
  \item{\dots}{additional parameters for the methods. For the RTUDocumentation
   object, this is the directory into which to write the separate Rd files.}
}
\value{
}
\author{
Duncan Temple Lang
}

\seealso{

}
\examples{
}
\keyword{documentation}
\keyword{programming}
\keyword{interface}
